fix(core): clear the two real MSVC warnings; audit the build flags
A clean MSVC build of this tree emitted 13 warnings in our own code. Two
classes were worth fixing and both are verified:

- C4805, twice in cascade.cpp: the shape-key hash folds three form-state
  functions on consecutive lines, two returning int and form_is_default
  returning bool, so one line mixed uint64_t ^= bool. Folding a bit into
  an FNV hash is intended, so this was not a bug — an explicit
  static_cast states the intent and matches its neighbours.
- C4389 in weva_c.cpp: Element::form_version() returns int64_t and was
  stored in a uint64_t field, so every comparison mixed signedness. It
  works because the counter only counts up, and would break quietly the
  day anything returned a negative sentinel. The field is int64_t now.

Left alone, with reasoning in AUDIT.md: eight C4190 warnings where
internal helpers sit inside the extern "C" block at weva_c.cpp:3504 and
return std::string / std::vector. Not a bug, but the ABI boundary is
untidy by accident of where the helpers were typed; moving them wants the
full gcc + clang-ASan run to land safely.

Verified on MSVC, which builds and tests this worktree even though WSL is
blocked in a worktree-isolated session: warnings 13 -> 8, core suite
505,882 checks / 0 failures before and after.

Also recorded: neither toolchain warns on narrowing conversions
(C4244/C4267 suppressed on MSVC, -Wconversion absent on gcc), which is
the class that guards the project's stated bit-identical-geometry
property; and nothing anywhere uses -Werror or /WX.

+++ godot-port/libweva/src/cascade.cpp
@@ -536,7 +536,7 @@ uint64_t CascadeEngine::try_compute_shape_key(const Element& e,
     uint64_t h = kFnvOffset;
     if (shape_key_folds_validity_) { h ^= form_validity_selector_state(e); h *= kFnvPrime; }
     if (shape_key_folds_range_) { h ^= form_range_selector_state(e); h *= kFnvPrime; }
-    if (shape_key_folds_default_) { h ^= form_is_default(e); h *= kFnvPrime; }
+    if (shape_key_folds_default_) { h ^= static_cast<uint64_t>(form_is_default(e)); h *= kFnvPrime; }
     if (!container_queries_.empty()) {
         h ^= container_provider_ ? container_provider_->version(e) : 0;
         h *= kFnvPrime;
@@ -584,7 +584,7 @@ uint64_t CascadeEngine::try_compute_shape_key(const Element& e,
     for (const Element* a = parent_el(e); a; a = parent_el(*a)) {
         if (shape_key_folds_validity_) { h ^= form_validity_selector_state(*a); h *= kFnvPrime; }
         if (shape_key_folds_range_) { h ^= form_range_selector_state(*a); h *= kFnvPrime; }
-        if (shape_key_folds_default_) { h ^= form_is_default(*a); h *= kFnvPrime; }
+        if (shape_key_folds_default_) { h ^= static_cast<uint64_t>(form_is_default(*a)); h *= kFnvPrime; }
         uint64_t anc = 0;
         anc ^= hash_str(a->tag_name());
         anc ^= hash_str(a->id()) * 257ULL;
